app/requirements.txt
discord==2.1.0
discord.py==2.1.0
PyNaCl==1.5.0
mysql-connector-python==8.0.31
sendgrid==6.9.7
google-cloud-logging==3.2.5